+++ validmind-toc/.gitignore
@@ -1,0 +1,6 @@
+NOTES.md
+out
+dist
+node_modules
+.vscode-test/
+# *.vsix
