--- vm-jupyter-toc/.gitignore
@@ -1,6 +1,0 @@
-NOTES.md
-out
-dist
-node_modules
-.vscode-test/
-# *.vsix
